Important bugfixes for shim-signed-common.postinst
As we're calling grub-install and specifying the target EFI arch, make
sure that we're using the *right* EFI arch. If we're on a mixed-mode
x86 system (32-bit UEFI, 64-bit CPU), that means that we need to be
specifying i386-efi here.

Also check to see if grub-efi has been configured to use
--force-extra-removable or --no-nvram. If we're calling grub-install
here, then it's important we follow the same rules or we may end up
breaking installations. :-/